+++ src/services/cache/EntitlementIdToSourceNameCacheService.ts
@@ -1,0 +1,16 @@
+import { IdentityNowClient } from "../IdentityNowClient";
+import { CacheService } from "./CacheService";
+
+/**
+ * Cache the mapping entitlement id-> Source name|Entitlement Name
+ */
+export class EntitlementIdToSourceNameCacheService extends CacheService<string>{
+    constructor(readonly client: IdentityNowClient) {
+        super(
+            async (id: string) => {
+                const entitlement = await client.getEntitlement(id);
+                return `${entitlement.source.name}|${entitlement.name}`;
+            }
+        );
+    }
+}